Tor everything.
@pretyflaco This :)
Not finished yet. The `torrc` file with bridges will be deleted when `_bootstrap.provision.sh` calls `/home/admin/config.scripts/internet.tor.sh on`.
But it is simple. Just need to concatenate the bridges backup file that is made automatically with the new torrc generated. This will restart Tor with bridges again and be safe.
I will need to test the whole build_sdcard.sh with this modifications, because I only tested the parts I added.