Update internet.tor.sh
making bridges being added if they exist, and not overwritten. This will serve for the build_sdcard.sh. But addind bridges could be a function to be called too from here.
The bridges from building the sd card will still be saved on a separate file as backup. But the optimal way is to make /etc/tor/torrc.d.